Don't attempt to backup files in .git
If you have `EDITOR=vim` and execute `git commit`, you'll be prompted
with a vim editor to add your commit message. Upon writing your changes,
a file is saved in .git, in a file named COMMIT_EDITMSG. While the file
gets copied to the backup directory, the file is also ignored due to it
existing in a directory named `.git`.

I wasn't able to identify what is ignoring that directory, but I suspect
it might be a safeguard in git itself, not to allow directories named
`.git`.

Signed-off-by: Beshoy Girgis <shoy@1ds.us>